Remove `register_user_scopes` from profile collection
Because scope collection is more of a data structure that users might use them selves. And allocating id's is more of a puffin specific thing. This was feedback from Traverse.
